itemsets: fix pre-selecting of default answers for value-mode itemsets.
i can still think of situations where problems could arise-- namely when trying to display the answer to the itemset question before it has been answered for the first time.
technically, the itemset should be initialized at the start of the form, and the Selection objects in the answer linked up to the appropriate choices, but this seems really hard!

--HG--
branch : itemset-dev